+++ TombEngine/Game/Lara/lara_helpers.h
@@ -33,6 +33,7 @@ bool HandleLaraVehicle(ItemInfo* item, CollisionInfo* coll);
 void HandlePlayerLean(ItemInfo* item, CollisionInfo* coll, short baseRate, short maxAngle);
 void HandlePlayerCrawlFlex(ItemInfo& item);
 void HandlePlayerFlyCheat(ItemInfo& item);
+void HandlePlayerExtraAnim(ItemInfo& item);
 void HandlePlayerWetnessDrips(ItemInfo& item);
 void HandlePlayerDiveBubbles(ItemInfo& item);
 void HandlePlayerAirBubbles(ItemInfo* item);