+++ TombEngine/Game/spotcam.h
@@ -4,7 +4,7 @@
 
 constexpr auto MAX_SPOTCAMS = 256;
 constexpr auto SPOTCAM_CINEMATIC_BARS_HEIGHT = 1.0f / 16;
-constexpr auto SPOTCAM_CINEMATIC_BARS_SPEED = 1.0f / FPS;
+constexpr auto SPOTCAM_CINEMATIC_BARS_SPEED = 1.0f;
 
 class Pose;