+++ TombEngine/Scripting/Internal/TEN/Flow/Settings/Settings.cpp
@@ -202,6 +202,11 @@ namespace TEN::Scripting
 			// @tfield bool enableInventory If false, inventory will not open.
 			"enableInventory", &GameplaySettings::EnableInventory,
 
+			/// Kill enemies which were poisoned by a crossbow poisoned ammo or by any other means. If disabled, enemy hit points will
+			// reach minimum but will never go to zero. This behaviour replicates original TR4 behaviour.
+			// @tfield bool killPoisonedEnemies If false, enemies won't be killed by poison.
+			"killPoisonedEnemies", &GameplaySettings::KillPoisonedEnemies,
+
 			/// Enable target occlusion by moveables and static meshes.
 			// @tfield bool targetObjectOcclusion If enabled, player won't be able to target enemies through moveables and static meshes.
 			"targetObjectOcclusion", &GameplaySettings::TargetObjectOcclusion);
@@ -426,6 +431,18 @@ namespace TEN::Scripting
 		/// Shadow text color.
 		// @tfield Color shadowTextColor A color used for drawing a shadow under any rendered text.
 		"shadowTextColor", &UISettings::ShadowTextColor,
+
+		/// Title logo center point position.
+		// @tfield Vec2 titleLogoPosition Center point of a title level logo position.
+		"titleLogoPosition", &UISettings::TitleLogoPosition,
+
+		/// Title logo scale.
+		// @tfield Vec2 titleLogoScale Title level logo scale.
+		"titleLogoScale", &UISettings::TitleLogoScale,
+
+		/// Title logo color.
+		// @tfield Color titleLogoColor Title level logo color.
+		"titleLogoColor", &UISettings::TitleLogoColor,
 			
 		/// Title menu position.
 		// @tfield Vec2 titleMenuPosition Title level menu position. Horizontal coordinate represents an alignment baseline,
@@ -438,7 +455,6 @@ namespace TEN::Scripting
 
 		/// Title menu alignment.
 		// @tfield Strings.DisplayStringOption titleMenuAlignment Specifies menu alignment.
-		//
 		// Can be set to @{Strings.DisplayStringOption.CENTER} or @{Strings.DisplayStringOption.RIGHT}.
 		// If set to `nil`, or set to any other value, menu will be aligned to the left side of the screen.
 		"titleMenuAlignment", &UISettings::TitleMenuAlignment);